Replace Highcharts with Plotly
Plotly (https://plot.ly/javascript/) does not have the licensing
problems discussed in #53.  Furthermore, it has more features and plot
styles, which will be useful later, like in #36.

In this commit I tried to keep the styling similar to the one with
Highlights.  I made small changes, that I consider improvements, like
making the plot adapt automatically to the window size -- including on
resize.