Do not magically stringify the declared parameter name
The parameter name is now required to be a string already.  There is no
fundamental reason why parameter names should be limited to identifier
names rules.  The fact that `NONIUS_DETAIL_UNIQUE_NAME` already exists
is convenient here.

This changes syntax for the user from:
```cpp
    NONIUS_PARAM(name, 42);
```

To:
```cpp
    NONIUS_PARAM("name", 42)
```

Note the semicolon is not needed anymore either.