Add parsing of parameters that should be tested at different sizes
The syntax is:
```
    -p <name>:<oper>:<init>:<delta>:<steps>
```

This tells that the benchmark should be run with `<steps>` different
values for parameter `<name>`.  Each value is computed as:
```
    next = last <oper> deta
```

Where `<oper>` may be `+` or `*` as for testing with linear or
exponential increments (the later is useful when checking logarithmic
algorithms or subtle non-linearities in the "constant factors" of an
implementation).