Simplify parsing of running parameters configuration
The old code was convoluted, with too many assertions and implicit
assumptions.  Instead of generating all the samples that we want to try
in the argument parsing, we delay this logic to a more appropriate place
later, without loss of information.

The only drawback of the new configuration representation is that it may
be less flexible, as the previous one allowed to easily add passing
explicitly all the various parmeters to try.  We can, though, easily add
this later by adding more options to the configuration.