Perform operations on parameters in their declared value
We always store parameters generically as `std::string`.  The user
implicitly suggests a concrete value type with the type of the default
value of the parameter.  We use that type when we need to do arithmetic
on the parameter and also when checking that it parses correctly when
passed from the command line.  We do this by casting the strings back to
the original `T` using the `param_spec` interface.

This might seem overkill.  Another approach could be to type erase the
concrete values once intially parsed from the command line, using a
`boost::any`-like type erasure class.  However that approach is less
flexible, because it requires the user to provide the specific concrete
type when calling `chronometer::param<T>()` and just a mistake in
signedness or whatever could be fatal.  Just parsing everything into and
back from strings is more forgiving on numeric types, which is the most
common case.

This commit adds a new `example7` that where using ranged parameter
would not work correctly previously.