Enable parameters that only support arithmetic partially
In the last commit, we broke parameters of types that do not support `+`
and `*` operators.  Some of these parameters might be useful, for
example `std::string`.

Now, parameters may or may not support these operations.  When they do
not support, the benchmarks will fail when we try to run the parameter
through a range with those operations.

Funny, we can do:
```
   ./bin/examples/example6 -p string:+:hola-mundo:-yeah:10
```

But we can't:
```
   ./bin/examples/example6 -p string:*:hola-mundo:-oh,no!:10
```