Run all benchmarks for a parameter setting in one round
This commit refactors a lot of the code added in this branch. The
original goal was to reverse how `go()` runs through various
parameter values.  Before it did like:

    for each benchmark
        for each parameter
            run-benchmark()

Now it is more like

    for each parameter
        for each benchmark
            run-benchmark()

This new structure should make writing reporters easier, since it better
fits how the data could be presented to the user in meaningful ways --
we'll see more of this in later commits.

A couple of more refactorings has sneaked in this commit:

- `param_map` and `param_registry` etc. are not in `detail::` anymore.
  This is to better fit other parts of the library (e.g. benchmark,
  reporter) that are not in `detail` and arguably at the same level of
  abstraction from the user API POV.

- A couple of functions now have a more "data driven" approach, trying
  to reuse previously computed data more often, as to improve the
  overall performance.