Add API for writing benchmarks that support reusable fixtures
The old API for benchmarks that have fixtures is like this:
```cpp
NONIUS_BENCHMARK([] (nonius::chronometer meter){
  // setup code
  m.measure([&] {
      // benchmark code
  })
});
```

That API is nice, but since it has to assume that the benchmark code can
always mutate the setup, the setup can not be reused across samples.
This is ok in general, but for benchmarks that have expensive setups
relative to their measured time, time might be wasted uneedesly.

Our alternative syntax is like:
```cpp
NONIUS_BENCHMARK([] (nonius::parameters meter){
  // immutable setup code
  return [=] {
      // benchmark code
  }
});
```

Or even further:
```cpp
NONIUS_BENCHMARK([] (nonius::parameters params){
  // immutable setup code
  return [=] (nonius::chronometer meter) {
      // mutable setup code
      meter.measure([&] {
          // benchmark code
      })
  }
});
```

The idea is that when the benchmark function can take a `parameters`
argument, it is assumed to be a factory that then returns the actual
benchmark function.  This factory function is called *only once* for a
given set of parameters (in `prepare(...)`) and the resulting benchmark
function is kept in the `execution_plan` and reused for collecting all
the samples.

Furthermore, this syntax may be more intuitive for people that need
parameters, but do not need a chronometer.