Prevent infinite loop in run_for_at_least
When trying to figure out how long to run a test to get a meaningful
value, we may fall into an infinite loop when the whole benchmark has
been optimized away.  We now try to detect this case by hard-limiting
the number of iterations we do in the loop.

This partly solves #52.  A complete solution involves properly
deoptimizing the return value of the benchmark function.