Fix passing both fixed parameters and a parameter run
Now all the logic of deciding which parameters to execute is contained
in `generate_params`.  We removed the "feature" that `parameters` looks
up default values.  This way the code we remove dependencies on global
state, and make the feature more testable.  `generate_params` now merges
the default values with the fixed values with the values generated for a
parameter run.