Make HTML reporter show summary page
The HTML reporter has been changed to generate multiple plots instead of
just one.  The plots may be choosed with a drop-down menu at the top.
The drop down menu is always focused such that the user may use the
keyboard arrows to quickly navigate the plots.

These are the plots that are generated:

**summary plot:**

  - If the benchmarks are run with only one set of parameters, it is a bar
    graph showing the mean and stddev of each benchmark.

  - If the benchmarks are run with against various parameters, it is a
    line graph showing the evolution of the benchmark results for the
    different parameters.  The graph is shown in logarithmic scale if
    when the parameters are generated using `*`.

**samples plot:**

   - For every set of parameters, a plot is generated depicting every
     sample taken for every benchmark.  This plot is like the one that
     was generated before this change.