Make params API more terse and type safe
Before, parameters were store as strings.  No type information about the
parameter was stored.  The user had to specify the type both at the
declaration site and at the usage site.  The conversion rules where
thouse of parsing with `>>` and `<<`, but some operations (e.g. `+` or
`*`) where applied in the initially declared type.

That was messy.

Instead, now declarations define a special type tag that encapsulates
name and concrete type.  Parameter values are parsed only once and then
carried around in a COW-enabled type-erased wrapper.  When the user
queries a parameter, they just pass the tag, and get a value casted in
the right type already.

Sehr gut.