Add mechanism to skip benchmarks
The idea is that in some cases, we might not want to run a benchmark.
The typical is example is benchmarking various algorithms with
increasing problem sizes (using *parameters*).  For high values of `N`,
we might want to skip those algorithms with poor complexity, but still
compare the other algorithms.

This is highlighted in the new `example8`, where various sorting
algorithms are tested.  When running it like this:
```
    bin/examples/example8 -p N:*:1:2:20 -r html
```
the following output is obtained:

    https://sinusoid.es/misc/nonius/skip-example8.html

The way has been implemented includes the following changes:

- Benchmarks continue running after one benchmark throws an exception.
  This means that one can skip the benchmark just by throwing an
  exception.  Also, this means that the system is more resilient now and
  it outputs all other available results even if a benchmark actually
  fails because of an error condition.

- A new kind of exception `skip_error` has been added to explicitly
  signal that we would like to skip this benchmark.  This exception may
  be thrown with `nonius::skip()` for convenience.

- The HTML reporter now takes into account that some benchmark runs may
  be missing results and generates sensible output in that case.