Add toggle to sort entries in html generated report
In some cases, specially in the bar graph summary view when many
benchmarks are present, it can be useful to sort benchmarks based on
their results.  This commit adds a checkbox labeled `sorted` to the top
bar --where the plot chooser is-- that enables sorting the benchmarks by
order.

Note that sorting the benchmarks changes the color and checkmarks of the
traces.  It might be better to preserve the initial ones, but I was too
lazy to do so--I believe that we would need to manually asign colors and
markers instead of relying on plotly auto-asigning them.

Here is an example of the results:

    https://sinusoid.es/misc/nonius/sorting-example8.html